+++ EdenSynth/libeden/source/synth/envelope/ExponentialGain.cpp
@@ -10,7 +10,7 @@
 
 namespace eden::synth::envelope
 {
-	void ExponentialGain::calculateGain(double sampleRate, std::chrono::milliseconds duration, float initialLevel, float finalLevel)
+	void ExponentialGain::calculateGain(float sampleRate, std::chrono::milliseconds duration, float initialLevel, float finalLevel)
 	{
 		/* 
 		 * This gain should be 'exponentially linear', which means that plot in dB scale should be linear.
